Add files via upload
This model was used to synthesize information relating to timber species of interest. Specifically, tabular data containing information on network costs when transporting timber from the point of harvest to a timber products mill using a road network. This information was used in RStudio to estimate the financial cost of moving timber.