Add files via upload
This code intakes data from the USFS - FIA EVALIDator, cleans, formats, and exports QGIS-ready tables relating to growth to removal, net growth, and net removal for pulpwood, bolts, and sawtimber for a variety of timber species present in Wisconsin. This code creates atlas and attribute tables and lookup tables to streamline the creation of maps within QGIS using the Atlas function.